indu/requirements.txt
Pillow
mock